+++ client/src/pages/Home.js
@@ -43,7 +43,7 @@ export default function Home() {
 
   const [buildingInfoData, setBuildingInfoData] = useState({});
   const [employeeInfoData, setEmployeeInfoData] = useState({});
-  const [employeeCommuteData, setEmployeeCommuteData] = useState({})
+  const [employeeCommuteData, setEmployeeCommuteData] = useState({});
 
   const handleBuildingInfoFormSubmit = (data) => {
     setBuildingInfoData(data);
@@ -51,6 +51,7 @@ export default function Home() {
 
   const handleEmployeeInfoFormSubmit = (data) => {
     setEmployeeInfoData(data);
+    getEmployeeCommuteInfo();
   };
 
   const switchPagesButton = (state) => {
@@ -83,10 +84,10 @@ export default function Home() {
   };
 
   const getEmployeeCommuteInfo = async () => {
-    const response = await fetch('http://localhost:8080/process-data', {
-      method: 'POST',
+    const response = await fetch("http://localhost:8080/process-data", {
+      method: "POST",
       headers: {
-        'Content-Type': 'application/json',
+        "Content-Type": "application/json",
       },
       body: JSON.stringify(employeeInfoData, null, 2),
     });
@@ -166,15 +167,15 @@ export default function Home() {
           {showEmployeeInfo && (
             <EmployeeInfo onFormSubmit={handleEmployeeInfoFormSubmit} />
           )}
-          {showEmployeeInfo && (
+          {/* {showEmployeeInfo && (
             <pre>{JSON.stringify(employeeInfoData, null, 2)}</pre>
-          )}
-          {showEmployeeInfo && (
+          )} */}
+          {/* {showEmployeeInfo && (
             <Button onClick={() => getEmployeeCommuteInfo()}>TEST</Button>
-          )}
-          {showEmployeeInfo && (
+          )} */}
+          {/* {showEmployeeInfo && (
             <pre>{JSON.stringify(employeeCommuteData, null, 2)}</pre>
-          )}
+          )} */}
           {showResults && <Results />}
           {showSimulation && <Simulation />}
         </Grid>
